Emitter now creates code blocks for functions.
Previously, each function would start off with a code block of size 8
and grow it over time. Now, code gets written into one single block in
emitter. When the function is done, the emitter slices the code and
hands it off to the particular function (except for __main__ which
gets a shallow copy of emitter's code).

This was added for two reasons:
* It's an optimization, as large functions won't cause a bunch of
  reallocs as the function's code scales upward from 8 slots.
* This will make closures easier, since my plan for them involves
  performing some rewriting on code. The rewriting will be easier if
  there's a large space to rewrite the code into.