Finally: Calls no longer 'damage' trees. This is super huge.
Up until now, when the emitter evaluated a call, said call had free
reign to adjust tree state. The one important thing that it did was
that it would go to trees of type tree_oo_access

'arg_start->result = arg_start->arg_start->result'.

If a tree needed varargs and didn't have any, it was patched in. Also,
if self needed to be set, an argument was replaced with the value of
self. This was becoming more awful over time.

Now, calls put their values into a buffer sort of like how patches
and pretty much everything in emitter gets a buffer with a starting
position. A new call state has been added and that's tossed around
and contains everything something might need to know about a call.

A bug was fixed in preparing __main__ where the intepreter would
have a potential crash if __main__'s bytecode was aligned to a power
of two and pretty wide (say, 128 spots big).