Refactor oo_access to be less weird, bugfix
Up until now, oo_access would either result in a defined function
being set as the result, or a property dumped into a storage.

This is actually a bug. A defined function cannot ever be returned as
the result because it's not loaded into a register. Attempting to
subscript assign a dot-loaded method crashes because of this.

This also includes a refactoring of begin_call so that it takes the
tree of the caller and figures out things from there. This should be
suitable for some new features I have in mind.

This refactoring should make it easier to do property getter/setter
functions later on.