Un-break jump patching.
It turns out that emit->patches should not get adjusted code buffer positions.
If it does, then it patches to the wrong place, since patching is done before
adjusting. A test has been added that crashes when the jumps were broke.

This was not caught because many, many tests operate in global space, instead
of within functions (there's often no need to make functions for quick tests).