At long last, closures have landed. Finally fixes #51.
Closures work pretty much...how you'd probably expect them to. You can close
over something in a higher scope, regardless of if it's from a class or
another function.

This is done by creating a copy of a function but which has a gc entry and a
closure container. Each shallow copy of a function gets a shared closure
container between them. When each shallow copy of a function that is using a
closure goes out of scope, that's one less ref for the container. Mentally, it
seems ok. I have the gc backing stuff up. Functions are all tagged as maybe
circular now, but that should not affect much.

Lambdas are able to refer to self using a closure (self is put in the closure
if it needs to be). Class methods can still be used via x::y but refer to
class parameters in upvalues (in such a case, 'self' takes a closure inside).

I've added a few tests for closures, and verified that they don't leak in
valgrind.