Symtab now flags classes that are primitive, VAL_IS_NOT_DEREFABLE.
This is part of trying to squeeze some more performance out of the vm. The
symtab is now responsible for tagging classes that are primitive (not
refcounted). This allows the vm to test if something should be derefed by
doing flags & VAL_IS_NOT_DEREFABLE which supercedes VAL_IS_NIL_OR_PROTECTED by
checking for either pyrotected, nil, or primitive flags.

Some operations in the vm were adjusted so that they set VAL_IS_PRIMITIVE
instead of taking away nility.

As far as performance, some rudimentary benchmarks that I ran showed that
integers and doubles are about the same, speed-wise. But strings and other
stuff which is refcountable runs MUCH faster.