Simplify: Only one flag is needed to prevent deref, don't set value...
There's no need to set a value when initially setting registers up.
Registers always start off as being nil, and the value is never used
if nil-ness is set upon it.

I've also fixed an issue with the list::append function, where it was
undoing VAL_IS_NIL. That was, in the pass, completely ok to do.
However, doing that to an integer is a big no-no, because integers
are not refcounted. Instead, lily_move_value is now used, which is
the right thing to do.