Register grows are now done by doubles. Fixes #114.
It's quite a bit faster to do growing by doubles to prevent doing a
bunch of grows for a tiny amount of size. I'm marking this as closed,
but there's still a fair amount of speed to squeeze out of the vm.
However, I want to focus on other things now.