Functions hold their import now, lily_vm_raise_seed.
This makes it possible for a function defined within a foreign loaded
module to raise an exception that may or may not be dynaloaded. If it
is dynaloaded, then the exception information is then tied to the
appropriate module. Nothing uses this right now, but it's something
that I think pg should have.

Part of getting this to work was making it so that symtab now puts
the module a thing was declared within into a function value instead
of just the path. This allows the vm to determine what module that a
function belongs to (and thus associate the dynaloaded exception with
that module.