Method dynaloads in the right scope, and POSTGRE!!! Fixes #100.
The first thing, method dynaloads being fixed, makes it so that
postgre's function prototypes dynaload classes if they need to. This
came up in the Conn class, which has query which looks like this:

query(Conn, string, list[string] => Result)

This posed a problem, because Result needed to be immediately
dynaloaded within the right scope. So that works now.

And then there's postgre. The basic api is there now:

Conn::new : Create a new connection
Conn::query : Run a query, replacing ? with values.
Result::fetchrow : Fetch the next row in the result set.

It's not much, but at least Lily can have some sort of a basic
conversation with postgre now! :)