Greatly improve return handling. Finally fixes #113.
Previously, a function that claimed to return a value absolutely had
to have a return at the end of it. This technically works, but is very
annoying. It also blocks one other feature (class-level functions for
enum classes), and thus I have an axe to grind with it.

if/elif/else, try+except, and match all now keep track of if all of
their branches actually exit. In the case of if/elif/else, there must
be an else clause (so it's not possible to do if a > 0 and a < 0).
With try+except, there needs to be a catch for Exception. If all
branches exit, then the block's exit information bubbles upward. Both
raising a value and returning are considered things that exit.

I've also added a dead code check after both return and raise, because
it seemed like a nice thing to add. try+except will now report a
problem if there is an except clause after an except for Exception (it
does not trap duplicate cases or silly ones though).