Fix a very subtle, very nasty ts corruption bug.
I'm extremely mad at myself for two reasons: One, I knew about this bug and I
didn't consider it important. Two, I spent two hours chasing down what turned
out to be a core-level problem.

In the beginning, the type system says that it has 0 generics seen, and 0 for
the ceiling. This is great: If a function doesn't take generics, then there's
no reason to allocate slots to save those generics.

There was an extremely subtle bug in emitter's call handling. It would call
for a ceiling raise, begin a call, then eval the call. But what would happen
if the ceiling was 0, and the call dynaloaded something with 1-2 generics like
list::append?

Corruption.

ts has something that has 1-2 generics, but they're in the ceiling. So they
get smashed. The type system is broken, and it becomes possible for A to
silently not be A. This is worse than bad.

So...raise the ceiling after beginning a call, so that the ceiling is raised
after any dynaload is done. Oh, and add a test. I never want this coming back.