Emitter: Create function values when they're needed...
...instead of creating them within the block and letting them linger
there. There's no reason for this, now that code is in one large chunk
and doled out when functions close.

This also fixes functions being tied wherein the tie would have the
wrong type. I'm not sure if that would be a bug or not (I don't think
ties go through normal assign). If it was, it's fixed. :)