Internal: A block's class_entry now bubbles up, fixing 2+ problems.
This was causing class methods to be created without marking that they
were part of a given class. But since this was only a convenience for
printing later (and not used for deciding if something was a class
method), it was easily missed.

For simplicity, emit->current_class is now gone. Class information is
tracked in blocks, and bubbles upward.

There are two problems that this corrects:
* A dynaload within a class method would interrupt the current class,
  causing anything after it to not inject 'self' as it should.

* An error raised in a non-::new class method would not have the class
  name printed before it, resulting in incorrect-seeming traceback.