Inject self for base methods + base methods from derived classes.
This is two things at once, but they're so intertwined that it would
be silly to have them come in separately.

First, if you had a base class method, then a derived class would not
inject self when calling it. This would cause it to fail, unless you
were to use a static call.

Second, a derived class could not call base methods without using
static access. It is now possible to do such.