Enum declarations no longer need commas between variant cases.
The braces make it obvious where the enum class begins and ends. It's
also very obvious where the types are because there are no
expression-like entities inside of an enum class (just types and
member declarations). Because of that, there's no ambiguity caused by
removing the commas, but it does make for a slightly terser syntax,
and I like that.