Prevent closing over vars in a certain ML case...
If a defined function has generics, it could have closed over a var
in another defined function with generics as well. In such a case, it
is possible to crash if the A of one define is solved as something
other than the A of another define.

I might add scoping to generics later, which would fix this. Until
then, it's being gated off to prevent a potential segfault.