F#'s awesome call piping (|>) has finally been added. Fixes #105.
There are some caveats to this. For one, you can't pipe to a lambda, because
Lily won't have inference information (this requires a considerable effort to
fix).

Piping has precedence under bitwise operations, but over comparison ops. I may
change this later, given a suitable reason to do so.

Piping will automatically work with dot-access like stuff. This means that you
can do ```10 |> [1,2,3].append```, and it works exactly as it should. If the
target is a class method, then injection of self will get triggered as if the
function has been called normally.