Finally, a boolean class. Fixes #125.
To begin with, this adds two new 'keywords': true, and false. These
are both mapped to yield back literals of a boolean type. This also
means that these values can't be assigned to (since they're literals).

Some string class api's have changed, notably the is* ones. They now
push back a boolean value, instead of an integer one.

Comparison operations: < <= == != >= > will now yield back a boolean
result, instead of an integer one.

I'll admit that boolean is a bit hacky in placement (it's written in
after string and bytestring internally). I'll likely fix this later
if/when I add a byte class.

Finally, the boolean class has two methods: to_i (conversion to an
integer), and to_s (conversion to a string). I've also updated show so
that it will display boolean values.