It's now possible to catch exceptions that include generics.
This allows things like ```except MyError[A] as e:```. I might regret
this later, but it sounds -really- cool now. :)

I've also done a slight change to instances. They used to have a
'true_class', but I've changed that to a 'true_type'. This means that
instances will be more accurate about what they hold. I did this
because it seemed sensible for implementing the better exception
catching.