Go back to having per-block loop starts.
The idea of having loop starts be per-emitter was great...except that
there's no place that loop starts are actually saved (using code_start
was actually a lie).

It worked, until you tried to do a nested for loop. The exit jump to
the outer loop would be set to the loop's code_start (likely zero). So
if you had a nested for loop within __main__, it would infinitely jump
back to the very start of __main__.

Yeah, that's pretty bad. So add a test to make sure this doesn't come
back again.