Fix a 'double wrap' oops with closures.
Previously, the emitter would take something like {|| x * x} and mark
it as an upvalue that wasn't closed over, twice. This would result in
the first x being properly attached to the upvalue x, and the second
not being attached to anything. Thus, {|| x * x} would always return
0, because the second x would not be wrapped.

This only would happen when seeing the same open upvalue twice. Doing
{||var v = x    x * x} would work just fine, making it even stranger.

But now {|| x * x} works, by assuming that any upvalue could be closed
over or not, and working accordingly.