printfmt/string::format allow %s on any and all values now.
This is a step toward what #129 is calling for, as it allows anything
to be converted into something string-like. However, it doesn't allow
custom to_s methods to be defined on a class.

With this, print now takes A, as does file::write. A test for basic
stringification has been added to the api tests.