Stop the weirdness with any/enum class values.
Since the beginning of their introduction (or at least as far as I can
remember back without looking through the logs), type any has worked
in a peculiar way. Values of type any start off as being nil, and are
initialized when they are first assigned a value. From then on, new
assignments will write into the any's inner value. Later on, enum
classes were added that were really just any's in disguise.

There is a problem with this. It is currently not possible to cache
multiple invocations of a foreign call's registers. Why? Because the
first run of the function will have any's and enum classes create new
values. The second and subsequent runs will use the old values and
write into places they shouldn't, causing sadness.

This is bad. So now, when something is going to be assigned to an enum
class or any, it has to be boxed first. I made a new opcode,
o_box_assign, that does this. As a bonus, lily_assign_value is simpler
now (it no longer has to do that check for any's/enum class values).

A few things needed to be hit in emitter for this to work. One nice
issue I found, is that list/hashes were asking to box things into type
any when they already were. Lambda eval now explicitly boxes variant
types, because there's no longer an automatic variant->enum promotion
anymore.