Simplify function dynaloading.
The part that describes the type of the function doesn't need to
specify that it's a function. That's already obvious (dyna_function).
The name isn't needed, since that's already provided. If anything,
having to provide the name twice leaves a potential issue in later if
they accidentally don't match.

The result is smaller dynaload strings.