Create lily_copy_value to copy values, and use it.
This takes advantage of enum classes and type any no longer
potentially triggering the gc during simple assignments. The new api
will create a copy of a value and give it a ref bump if necessary.

Builtin list+tuple creation uses this, as well as list::fill. Based on
the above information, some other adjustments were made to both
things (such as not repeatedly setting the list count on each pass).