Use type_matchup for oo and prop assigns. It fixes a nasty bug...
There was an issue where assigning a variant type to a property or
through oo assign would cause it to not be boxed as an enum. This is
really bad, because bare variants are not supposed to happen. This
could result in assignments failing (mentioning the bare variants), or
in crashing from trying to show the variant values.