Generic solving now fully understands co/contravariance.
It may not seem like a change, because bare generic types are
unwrapped when doing argument processing. Because of that, bare
generics like A already benefitted from co/contra variance. But if you
passed something more complicated like...

    define f[A](a: A, f: function(A => A)) { ... }

then the A's that are function parameters would not understand that
variance and demand exact matching.