Traceback has been redone, and also turned into a list[string].
The new format is much easier to deal with: There are a wealth of
string functions, but nothing works on tuple[string, string, integer].

I've also redone the actual format, so it looks like a mix of Lua and
Ruby's traceback. From Lua, I took the idea of including C calls into
the stack (I like to know everything), and from Ruby I took the idea
of having : as a delimiter between values.