Fix a long standing bug with nested single-line if's.
Consider this:

if 1: if 1: 1 else: 1 else: 1

The if branch has an inner if branch that is also a single line. Each
of them has a separate else branch.

The parser used to think that both else blocks should go against the
inner if block, and then complain that was wrong. Now it correctly
understands to immediately finish after an else.