parser: If a single-line try excepts for all, it's done.
It's the same thing as the last patch, but for try+except instead of
if branches. If there's an except Exception clause, then there's
nothing more of interest for that branch.

The existing test, unreachable_except.lly, was changed so that it
would still fail in the same way. It just needed to be made multi-line
so that the branch wouldn't close early.