At last, types are finally bucketed by the class they're made in.
Previously, the symtab had a 'root_type' that every type ever created
was linked through. This was initially pretty good, because the idea
is that you have a one-stop shop for all type stuff.

However, it isn't that great at scaling. One problem is that when
there are a lot of types, a lot of them are not at all relevant.
Trying to build a list of something would mean wading through all the
function types created so far, the hash types, the tuple types, etc.

Now, when a type is created, it's placed into the class that it's
created in through the 'all_subtypes' field of the class struct. This
means that searching is now faster because it's now only done over
relevant types.

I've also cleaned up how generics are done a bit. The lookup_generic
function in symtab was HORRIBLE, but looks quite a bit better now.