Introduce new module: tm. This one builds types.
Up until now, type creation has been very, very fragmented.

Parser:
    New types are created through having a type stack and shoving
    things directly into the type stack. Symtab would be asked to do
    the actual build.

Emitter:
    Make ts carry a ceiling and shove extra types into the ceiling.
    Tell ts to build something based on the ceiling.

ts:
    Store types in the ceiling manually and ask symtab to do the
    build. This one is understandable, because resolve has to be fast.

Symtab:
    Actually do type building. However, symtab does not actually have
    a type stack or use these types. It just builds them using a sent
    stack, a size, an offset, and other things.
    As one might expect, this led to "oops" issues of too much/little.

Now, all type building (except for symtab building generics and
defaults) has been unified within a new module called type maker. No
more parser fiddling with pushing and popping a type stack. No more
emitter storing ceiling types.

The new system is a lot safer, because a module like parser or emitter
will insert the number of subtypes it needs and call for a type make
with just how many to take off the top. So "oops" issues are not a
thing of the past, but certainly much harder to do.

As a sign of stability: I completely rewrote how enums are build from
variants, and the resulting code was successful the first time. The
same can be said of lambdas. This is a very good module.

There is still work to be done in this area: Parser's dynaloading
needs a similar unification so that symtab can be freed of having to
make default types. I also need to determine the future of generic
type lookup, which is in symtab but does not seem to belong.

For now, this module represents a rethinking of how type building
should be, and it is a very good thing.