Parser: Use imports+classes being items to fixup dynaload_function.
Previously, dynaload_function would call parse_prototype. But it only
had 5 lines or so and was the only caller of parse_prototype. That's
the first problem.

Another problem is that not being able to unify imports and classes
meant that there were two very similar but not quite so functions to
create a dynaloaded method or dynaloaded toplevel function.

I've merged dynaload_function with parse_prototype, and merged the two
arguments into one lily_item, making the code simpler and cleaner. The
two emit functions have been similarly merged into a single one that
is much cleaner (I also removed a wrong comment).

Unfortunately, this means adding some more prototypes to the mess
already at the top of parser. Sorry about that.