ast: Boot the pool with 4 trees, and don't take a parameter.
I had ast pool take a parameter a long long time ago because I had
this thought of various bits of the interpreter like that being
configurable. This ended up being useful initially, because I think I
may have had an array of trees a long time ago. Also, 8 was chosen
because trees used to also function as save nodes (instead of having
custom save nodes).

Because of the advancements in Lily's core design (saves are now by
themselves, trees are linked lists), I think 4 is a better choice
also. I've also been reluctant to allow tuning of internal parameters.
But if I do, it'll be through options and not directly passed
parameters.