ts shouldn't do raw make anymore.
This patch puts ts a little further away from using a ceiling. I
eventually want ts to stop using ceilings entirely, because then ts
raise will grow less each time (since it will not need to account for
a ceiling). That, in turn, will mean smaller grows and all around
faster ts speed (though only by a few hairs).

This makes it so that ts uses two new tm apis to add a type without a
size check, and to do a giant size check all at once. This is vital
for resolve, because resolve is used in the vm (and thus has to be
extra fast).

This patch also stops resolve from using the aforementioned ceiling
and having to adjust the ceiling before/after calculation.