Create lily_buffers, and the first buffer (uint16_t).
I have to give a shoutout to Wren because that's where I got this idea
from. I'm a bit ashamed that I didn't think of it sooner, because I'm
starting to get tired of all these buffer growing functions that are
being littered around.

Sadly, this does not include fixing other buffers in emitter. The
reason for that is that I want to fix how some of those buffers are
used (match should probably use jumps like exceptions instead of a
single glob, closures could probably use that padding in vars, and
likely more).

My current plan is to go back to those buffers and fix them up after
doing more type system work (again).