eval_call now sends down ?-solved types.
Ever since generics were added, the common thing to do was to figure
out if the current function was or was not holding solved generics. It
was a big problem because the handling was confusing and thus often
wrong. Wrong handling of the solving bit means either poor inference,
or wrong inference (double solving).

Now that there is a ? type, it can be safely substituted and assigned
over. So instead, fill in unsolved generics with the ? type and let
that be set to NULL if that's what needs to be done. The result is
that emitter's inference code (and parser's lambda inference code) is
a lot more straightforward.

With this change, inference will be better throughout the interpreter
because it's a LOT harder to be incorrect with inference.