apache: Add three server functions: write, escape, and write_raw.
This is the start of the type safety stuff for the server. One thing to
note about this is that server::write is the first function which checks
if a value is a literal and denies it if it isn't. The reason for this is
to allow server::write for fast writing, and server::write_raw +
server::escape for writing out escaped stuff.