Create the Tainted class to represent untrusted data.
Since postgre and server are both going to need this, it's going to need
to be in the core. For now, I'm going to put Tainted into core directly.
Some time in the future, Tainted will be dynaloaded (and hopefully
Exception as well).

The Tainted class represents a container where data goes in and can only
be retrieved through a sanitizing function. The sanitizing function is not
particular, however: One may have a Tainted[string] and use string::to_i
to not only retrieve the data, but have it as an integer (unless there is
an exception thrown).