Fix a VERY subtle bug in try block popping and return.
The problem is that the calculation for return block popping was only
going up to the current loop block, instead of the current define
block. This meant that it was possible for some try blocks to not get
popped from an early return.

A test has been added that crashes without the fix (because the try
block is used and then jumps to the wrong area).