vm: Interior exception handling should handle vm_list unwinding.
The old way was something that was easy to be forgotten (functions were
expected to set a catch and rollback vm_list pos on error). The new
way is simpler, because functions don't need to set a jump link when
calling into the interpreter. Instead, they simply call into the
interpreter and don't have to worry about fixing vm_list->pos if there
is an exception that passes through them.