Unify common code and fix a VERY subtle bug in new list API.
The list being iterated over was getting the tag (possibly a second
tag!), instead of the newly-created list.

As for the unification, well, list iteration functions now copy the
value into vm_list and then shallow move it out later on. Before
this, select and reject were shallow moving it into vm_list and then
deep copying it later on.