Don't re-use any/enum values, prevent variant optimizing in assign.
While trying to fix the circularity bug, I noticed that box assign in
vm was always creating a new any value and tagging it. That seemed
wasteful, but the real problem is that it was never deref-ing old
values. It was trivial to assign a refcounted value over and over
again to prevent it from being destroyed.

Fixing it wasn't easy though. The obvious technique was to not create
new boxed values and instead use the old one. Except...sometimes the
emitter likes to re-use storages.

So what happens if you have a var assigned to an Option[integer], and
that Option[integer] has a copy in a storage? Bad things. The next
time that something grabs the storage, it's going to overwrite what is
inside of the contents. So now the var's value is changed, as if by
some sort of magic.

I've thus made a decision: Enums, variants, and type any may not have
storages reused. Once a function grabs an enum/variant/type any
storage, that storage sticks until the function is done. No reuse at
all.

Another issue was the semantics of enum types. I decided to disable
optimizing variant values being optimized, because it meant that one
could assign to a parameter that is an enum and have the parameter's
inner value mutated. That just didn't sit right, so I took that out
as well.