Properly mark circularity...sort of (Fixes #94).
This is a hard problem to solve, because of variance, and because a
function can have a container of the current class without the current
class being necessarily circular.

This fixes the interpreter to assume a class needs a gc marker if it
contains a container of itself anywhere inside of itself, or has
something inside that needs a marker. This isn't a great fix, but it
is a lot better than what was there before, where enums were always
circular and classes...were busted.

It's overly aggressive, but at least it's a big step in the right
direction.