Fix an oops in type maker with variant types.
It turns out that the patch that added ? solving kinda fixed the issue
that existed with partial solving with variants. But...there was
another problem. When trying to build the result for a variant, the
wrong value was being used. That wrong value would end up causing a
segfault.

This fixes that, and therefore unleashes partial unification which
somehow just falls into place. Neat!