Deleted about 100 tests, then made a handful of new ones.
A few of the tests were back from when Lily would segfault if you so
much as looked at it funny. Now, Lily is a lot more solid, and some
of the older tests just never failed. Some of the tests had also been
put in as safety checks (You completely broke X feature), such as the
optargs_basic.lly test.

What's been bothering me lately (and what prompted this test), is
that too many passing tests are crashing when there is a problem. So
many that I have a sea of panics from 30-40 tests when a single thing
is broken. That isn't helping!

So I've consolidated a lot of tests. Some of the useless ones are now
completely gone (looking at you, andor.lly), and some of them have
been merged into tests that are ~100 lines. Call piping, for example,
had all of the tests get merged together because piping is a really
easy feature that just builds upon other features.