Okay, I was wrong. Make any and enums pass by value, not ref.
With this change, the folding that was done for type any and enums
can be done once more. I decided this after seeing that a function
that took an any could assign it to a different value and that new
value would stick.

Now if you send an enum/any to a function, the function can assign to
it without changing the original. I dislike far-reaching magic.