I finally solved that and_then problem. Fixes #120.
The proper solution is for lambda that have closed over an unsolved type to
put that information aside. Generic register solving then comes in and uses
the upvalues as a means of fixing the types. Without doing so, the B type of
and_then is unknown and defaults to any (that's bad).