Add list::pop to remove items from the end of a list.
I fully realize that the way that I'm allocating and free-ing list
elements is likely not the best one. However, I also understand that
trying to make a super optimized list type is something beyond the
scope of my knowledge. So, instead of trying something likely wrong,
I'll stick to simple and improve it later.

Also, I'm aware that list::append is due for a rename. It will happen
shortly.