The internal hash api is now cleaner, and where it should be.
It was previously in vm because...I can't really remember why. The
other problem is that the hash api required explicitly creating a
siphash using the vm's sipkey. All sorts of potential for doing
something the wrong way.

Now, there are three internal hash api functions which take the vm and
the vm handles the siphash part. It's harder to do it wrong now, and
that's a good thing.