Forbid adding/removing hash elements during iteration.
Stuff breaks when it's allowed. I could allow removal, but then the
iteration functions need to check for removal. What if it's just one
item ahead or five items ahead? It's a mess that I'm not quite willing
to get into just yet, so just forbid it entirely for now.