Provide lily_release_jump to release unused jump.
This wraps over a single line of code, and thus might seem silly.
However, I'd rather have a well-defined api function over 'make sure
to manually adjust the jump if it's not used'. The latter is rather
yucky-sounding. The former is also a lot clearer, answering the
question of 'okay, what happens if the jump isn't used?'