Closure jumping is now finally fixed.
This is a subtle problem that has existed since closures were first
introduced. It happens when there is a value that is closed over is
used again in the original scope in some sort of condition. It works
like this:

Suppose that a closure is being rewritten so that closure reads/writes
are injected where they need to be. Currently, there are 40 slots of
extra reads/writes. So a jump that previously went to 100 would need
to go to 140, right? Well, not necessarily.

For jumps that go backwards, yes. For jumps that go forward, no. What
if, between 80 and 100 there were more reads/writes injected in? Now,
the jump that was at 100 is going to be at 150. But, somewhere that
wanted to reach 150 has 140 written in.

The result is that the vm jumps to wrong places, and executes bad
opcodes. In the test that I've added, not doing this patching causes
a segfault because of those instructions.