Redo for loops (and fix a semi-related bug too).
Previously, Lily's for loop would try to auto-detect if it should do
+1 or -1 when looping. This actually turns out to be a REALLY BAD
idea. Sometimes I want to not loop if there are not going to be
values to loop over (instead of a loop in reverse). One example that
I was trying to do hit this wall.

Also, loops now do a reverse step before starting a loop, instead of
skipping the check the first time. This means that loops like 10...0
would run before, but will no longer do so. They should not have
ever worked.

The bug was caused by grabbing a local storage to hold the counter
when doing a global for. This is a bad idea, because the storage is
almost certainly going to be snagged by something as a storage
target. This was not previously evident, because of some internal
semantics of how for loops work.

I am aware that this leaves for loops with an unused slot that gets
emitted. However, I will not fix this until the next release, as I
do not want to break anything, and I have plans for redoing a fair
amount of emitting anyway. I'll simply include the reduction in the
emitting redoing.