Fix a really strange bug (`VList(VList()VList())` crashes).
The problem is that parser was allowing tk_label to escape but not
checking when. So an escape could be done during a nested expression
which is incorrect. To further complicate the matters, the VList
(assume it's a variant that takes varargs), would get extremely
confused and...infinitely recurse into eval. By fixing the label
case, the improper behavior and crash become (I hope) impossible.

A test has been added to make sure the problem never returns.