Add 'try/' with a few examples. Fixes #119?
These examples capture the bare basics of Lily. It's a language that
started out more C-like, but has evolved to rely on type inference and
functional ideas. These examples aren't that great though. I've held
off on doing more for a few reasons:

* Now that %s handles anything within printfmt, I want interpolation
  and I want printfmt to die.

* Option needs to be baked into the core. This will change, at least,
  string::to_i, which currently throws on error.

* Having to write @<name> = <name> in class init is annoying and in
  a big need of fixing. It's shameful.

* No containers support slicing, limiting examples a lot.

* No list::join, which hurts a lot.

Most importantly:

* This release has taken too long, and I want it to be over.

So, with all of that in mind, I'm calling this done. I'll contribute
more to these examples within the next release.