Emitter: Move empty list+hash building to a different function.
Building an empty list/hash is really annoying, so contain it to a
single function. This way, list and hash build can focus on just what
they do best.

Oh, and put the code that does the error for a bad key into one place
too, to simplify things some more.