Emitter: Variants now push back a ?-padded enum type.
This also includes setting a self type for enums and classes, because
that's an easy way of getting a ? solved type (just ask ts to solve
it). This type is now used by call for solving generics. This has a
nice side-effect of no longer defaulting a variant early if that isn't
necessary. Yay better type information!