Add lily_ts_unify. Static lists now require a consistent bottom type.
lily_ts_unify is a new api that will run the type engine to check for
equality, but also build a common type as it goes.

To make sure that it works, I've decided to use it to simplify the
building of static lists. Henceforth, there will no longer be a thing
called 'default to any' when there is a list of an uncommon type. In
order to get such a thing, one must use a typecast to coerce the data
into being what you want. It may not work though.

This is half of the unification that #140 asks for. However, it
largely relates to #148, as it provides a combination for variant
values that has knowledge of the padded types that are being made.

Some tests needed to be altered for this, mostly because a list of
inconsistent values no longer becomes a list[any]. Instead, one must
ask for that using a typecast.