Create partial_eval to handle eval when the target might be a variant.
This combines what list build and call arg handling were doing to get
the result when it may/may not be a variant. Keep in mind that this
code is severely flawed (does not handle parentheses, and piping is
totally not working still). However, encapsulating this in one area is
a good starting point!